Trying to implement MVC
